+++ src/main/java/com/georgewoskob/wav/WavTransformer.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 
 public class WavTransformer {
 
-    public void writeTextToWave(int[] wavArray, File outputWavFile) throws IOException, WavFileException {
+    public void writeTextToWave(double[] wavArray, File outputWavFile) throws IOException, WavFileException {
         WavFile wavFile = WavFile.newWavFile(outputWavFile, 1, wavArray.length, 16, 44100);
 
         wavFile.writeFrames(wavArray, wavArray.length);
@@ -13,11 +13,11 @@ public class WavTransformer {
         wavFile.close();
     }
 
-    public int[] wavToArray(File waveFile) throws IOException, WavFileException {
+    public double[] wavToArray(File waveFile) throws IOException, WavFileException {
         WavFile inputWavFile = WavFile.openWavFile(waveFile);
         inputWavFile.display();
         int numFrames = safeLongToInt(inputWavFile.getNumFrames());
-        int[][] buffer = new int[2][numFrames];
+        double[][] buffer = new double[2][numFrames];
 
         inputWavFile.readFrames(buffer, (int) numFrames);
